requirements.txt
Flask==3.0.0
PyPDF2==2.12.1
langchain==0.2.1
langchain_community==0.2.1
boto3==1.34.141
werkzeug==3.0.1
faiss-cpu==1.8.0
python-dotenv==1.0.1
sentence-transformers==2.5.1
pymongo==4.6.2
langchain_mongodb==0.1.6